Wrong locaton for syntax error
Syntax errors inside function/modules get reported at the wrong location. Most often at the first line of the function where the syntax error happens. For example take a look at ```file_input()``` it passes the module as AST node to the error function.
Looks like this if inside ```syntax_error_dbg()``` is the reason, I'm wondering if the condition is just swapped or if the issue has to be handled in a different place.